+++ spec/integrations/rails_4_2/test.sh
@@ -14,7 +14,7 @@ bundle --version
 bundle install --quiet --gemfile="$DIR/Gemfile" --retry 2 --path vendor/bundle
 
 echo '-- cypress install'
-bundle exec ./bin/rails g cypress_dev:install --cypress_folder=test/cypress
+bundle exec ./bin/rails g cypress_dev:install
 
 echo '-- start rails server'
 # make sure the server is not running
@@ -24,8 +24,8 @@ bundle exec ./bin/rails server -p 5002 -e test &
 sleep 2 # give rails a chance to start up correctly
 
 echo '-- cypress run'
-cp -fv ../cypress.json test/
-cd test
+cp -fv ../cypress.json spec/
+cd spec
 yarn run cypress run --record
 
 echo '-- stop rails server'