+++ spec/integrations/rails_5_2/app/models/post.rb
@@ -1,13 +1,2 @@
-class Post < OpenStruct
-  def self.create(attributes)
-    @all ||= []
-    post = new(attributes)
-    @all << post
-    attributes['all'] = @all.index(post)
-  end
-
-  def self.all
-    @all ||= []
-    @all
-  end
+class Post < ApplicationRecord
 end