+++ spec/integrations/rails_5_2/config/routes.rb
@@ -1,4 +1,5 @@
 Rails.application.routes.draw do
-  root 'welcome#index'
+  resources :posts
+  root 'posts#index'
   # For details on the DSL available within this file, see http://guides.rubyonrails.org/routing.html
 end