+++ spec/integrations/rails_5_2/db/migrate/20180621085832_create_posts.rb
@@ -1,0 +1,11 @@
+class CreatePosts < ActiveRecord::Migration[5.2]
+  def change
+    create_table :posts do |t|
+      t.string :title
+      t.text :body
+      t.boolean :published
+
+      t.timestamps
+    end
+  end
+end